A dedicated REST API resource for getting the current user.
Getting the current user used to be under a REST API resource /api/users,
which intuitively sounds more like a resource to list all users in the
system.

There isn't currently a use case in the front-end that requires getting
a list of all users.
