+++ ScampApi/wwwroot/App/Scripts/homeSvc.js
@@ -2,7 +2,7 @@
 angular.module('scamp')
 .factory('homeSvc', ['$http', function ($http) {
 
-    var apiPath = '/api/users';
+    var apiPath = '/api/user';
 
     return {        
         getUserProfile: function () {
